requirements.txt
awscrt==0.23.5
awsiotsdk==1.22.1

